# fix(perf): convert traverseDoc from recursion to stack (#4776)