@sompylasar 2018-07-16
### Before
![screen shot 2018-07-15 at 7 20 31 pm](https://user-images.githubusercontent.com/498274/42741449-33706e78-8867-11e8-9249-e8f0c21d7602.png)

### After
![screen shot 2018-07-15 at 7 20 36 pm](https://user-images.githubusercontent.com/498274/42741450-3387cf96-8867-11e8-9cc4-2433b973a381.png)