@sompylasar 2018-07-16
Note that here the performance is measured on the built code `./dist/bin-prettier.js` (in contrast to the previous improvement #4790 that was measured on the source code `./bin/prettier.js`).