@sompylasar 2018-07-16
Using `benchmark`, the difference is almost unnoticeable (also it's flaky, ranging from 126ms to 133ms):

Before (benchmark)
```
cat ../LspLanguageService.js | NODE_ENV=production node ./dist/bin-prettier.js --stdin-filepath LspLanguageService.js --loglevel debug --debug-benchmark > /dev/null
[debug] normalized argv: {"color":true,"debug-benchmark":true,"editorconfig":true,"stdin-filepath":"LspLanguageService.js","loglevel":"debug","plugin-search-dir":[],"plugin":[],"ignore-path":".prettierignore","debug-repeat":0,"config-precedence":"cli-override","_":[]}
[debug] resolve config from '/Users/ivanbabak/_sompylasar/_github/prettier-2/LspLanguageService.js'
[debug] loaded options `null`
[debug] applied config-precedence (cli-override): {"filepath":"LspLanguageService.js"}
[debug] '--debug-benchmark' option found, measuring formatWithCursor with 'benchmark' module.
[debug] '--debug-benchmark' measurements for formatWithCursor: {
[debug]   "benchmark": "format x 7.96 ops/sec ±9.11% (25 runs sampled)",
[debug]   "hz": 7.962900469589217,
[debug]   "ms": 125.58238092000002
[debug] }
```

After (benchmark):
```
cat ../LspLanguageService.js | NODE_ENV=production node ./dist/bin-prettier.js --stdin-filepath LspLanguageService.js --loglevel debug --debug-benchmark > /dev/null
[debug] normalized argv: {"color":true,"debug-benchmark":true,"editorconfig":true,"stdin-filepath":"LspLanguageService.js","loglevel":"debug","plugin-search-dir":[],"plugin":[],"ignore-path":".prettierignore","debug-repeat":0,"config-precedence":"cli-override","_":[]}
[debug] resolve config from '/Users/ivanbabak/_sompylasar/_github/prettier/LspLanguageService.js'
[debug] loaded options `null`
[debug] applied config-precedence (cli-override): {"filepath":"LspLanguageService.js"}
[debug] '--debug-benchmark' option found, measuring formatWithCursor with 'benchmark' module.
[debug] '--debug-benchmark' measurements for formatWithCursor: {
[debug]   "benchmark": "format x 7.98 ops/sec ±10.11% (25 runs sampled)",
[debug]   "hz": 7.976878149522964,
[debug]   "ms": 125.36232612
[debug] }
```